backend/requirements.txt
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncer==0.0.7
attrs==24.2.0
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
fastapi==0.112.4
filetype==1.2.0
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
Lazify==0.4.0
literalai==0.0.623
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.13.3
nest-asyncio==1.6.0
numpy==2.2.0
openai==1.57.1
opentelemetry-api==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-exporter-otlp-proto-http==1.28.0
opentelemetry-instrumentation==0.49b0
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
packaging==23.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-slugify==8.0.4
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
shapely==2.0.6
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.40.0
syncer==2.0.3
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.27.0
urllib3==2.2.3
uvicorn==0.25.0
watchdog==6.0.0
watchfiles==0.20.0
webencodings==0.5.1
Werkzeug==3.1.3
Wikipedia-API==0.7.1
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.2
